chore: moving from shared to me
